backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
google-api-python-client
fpdf
crewai
